factored out rings and added overlay tests
