+++ o/detect.py
@@ -2,7 +2,8 @@ from enum import Enum
 from PIL import Image, ImageFilter
 from profilehooks import timecall
 
-from ring import SimpleRing, HoughTransformRing
+from simple_ring import SimpleRing
+from hough_transform_ring import HoughTransformRing
 
 
 class DetectionStrategy(Enum):
